refactor: CRUD 별 서비스를 하나로 통합
