--- src/test/java/roomescape/service/reservationtime/ReservationTimeFindServiceTest.java
@@ -1,42 +1,0 @@
-package roomescape.service.reservationtime;
-
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
-import org.springframework.jdbc.core.JdbcTemplate;
-import roomescape.domain.ReservationTime;
-import roomescape.repository.ReservationTimeRepository;
-
-import java.time.LocalDate;
-import java.time.LocalTime;
-import java.util.Map;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-@JdbcTest
-class ReservationTimeFindServiceTest {
-
-    private JdbcTemplate jdbcTemplate;
-    private ReservationTimeFindService reservationTimeFindService;
-
-    @Autowired
-    public ReservationTimeFindServiceTest(JdbcTemplate jdbcTemplate) {
-        this.jdbcTemplate = jdbcTemplate;
-        reservationTimeFindService = new ReservationTimeFindService(
-                new ReservationTimeRepository(jdbcTemplate)
-        );
-    }
-
-    @Test
-    @DisplayName("날짜와 테마가 주어지면 각 시간의 예약 여부를 구한다.")
-    void findAvailabilityByDateAndTheme() {
-        LocalDate date = LocalDate.now().plusDays(1L);
-
-        assertThat(reservationTimeFindService.findIsBooked(date, 1L))
-                .isEqualTo(Map.of(
-                        new ReservationTime(1L, LocalTime.of(10, 0)), true,
-                        new ReservationTime(2L, LocalTime.of(11, 0)), false
-                ));
-    }
-}
